database/requirements.txt
psycopg2-binary
pgvector
python-dotenv
